+++ config/initializers/contour.rb
@@ -36,13 +36,6 @@ Contour.setup do |config|
       name: 'Sources', display: 'signed_in', path: 'sources_path', position: 'left', condition: 'current_user.system_admin?'
     },
     {
-      name: '@source.name', eval: true, display: 'signed_in', path: 'source_path(@source)', position: 'left',
-      condition: '@source and not @source.new_record?',
-      links: [{ name: 'Rules', path: 'source_rules_path(@source)' },
-              { name: 'File Types', path: 'source_file_types_path(source_id: @source.id)' },
-              { name: 'Joins', path: 'source_joins_path(@source)' }]
-    },
-    {
       name: 'Users', display: 'signed_in', path: 'users_path', position: 'left', condition: 'current_user.system_admin?'
     }
   ]