+++ config/routes.rb
@@ -76,12 +76,10 @@ SleepPortal::Application.routes.draw do
     member do
       post :auto_map
       get :table_columns
-      get :table_columns_for_select
       get :download_file
       post :remove_all_mappings
     end
 
-    resources :joins
     resources :rules
   end